hostilefork.json => now runs relative to cwd
As a new rule for Ren-C as a whole, scripts don't change the working
directory when invoked from the command line.  Instead, the feature
of finding resources relative to the script location is done by using
TAG! values as the argument to DO, LOAD, IMPORT:

https://github.com/metaeducation/rebol-issues/issues/2374#issuecomment-912266377

So long as the note needs changing, this goes ahead and mentions
that the whitespace interpreter is being used as a testbed for the
UPARSE parser that is written entirely in interpreted usermode
code, and is thus glacially slow compared to native code.  The
native code rewrite is being held off until a working stream
parsing design has been achieved.